* Added outfits/addons (not finished) * Support precompiled headers in devcpp * Change to 0.6.0 version number * Use _t types in networkmessage

+++ otserv/trunk/protocol.h
@@ -45,7 +45,7 @@ public:
 	virtual ~Protocol();
 	
 	void setPlayer(Player* p);
-	virtual void setKey(const unsigned long* key) = 0;
+	virtual void setKey(const uint32_t* key) = 0;
 	unsigned long getIP() const;
 	
 	virtual bool CanSee(int x, int y, int z) const = 0;
